+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "react-waypoint",
-  "version": "2.0.3",
+  "version": "3.0.0",
   "description": "A React component to execute a function whenever you scroll to an element.",
-  "main": "build/npm/waypoint.js",
+  "main": "build/npm/index.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/brigade/react-waypoint.git"
@@ -10,7 +10,7 @@
   "homepage": "https://github.com/brigade/react-waypoint",
   "bugs": "https://github.com/brigade/react-waypoint/issues",
   "scripts": {
-    "build-npm": "rimraf build/npm && mkdirp build/npm && babel src/waypoint.jsx --out-file build/npm/waypoint.js",
+    "build-npm": "rimraf build/npm && mkdirp build/npm && babel src/ -d build/npm/",
     "lint": "eslint . --ext .js,.jsx",
     "testonly": "karma start",
     "test": "npm run lint && npm run testonly",
@@ -27,7 +27,7 @@
     "babel-cli": "^6.0.0",
     "babel-core": "^6.0.0",
     "babel-loader": "^6.0.0",
-    "babel-plugin-add-module-exports": "^0.1.2",
+    "babel-plugin-add-module-exports": "^0.2.1",
     "babel-plugin-transform-react-jsx": "^6.0.0",
     "babel-preset-es2015": "^6.0.0",
     "eslint": "^2.11.1",
@@ -40,10 +40,10 @@
     "karma-firefox-launcher": "^0.1.6",
     "karma-jasmine": "^0.3.6",
     "karma-webpack": "^1.7.0",
+    "mkdirp": "^0.5.1",
     "react": "^15.0.0",
     "react-dom": "^15.0.0",
     "rimraf": "^2.5.2",
-    "mkdirp": "^0.5.1",
     "webpack": "^1.5.3"
   },
   "keywords": [
