+++ src/index.jsx
@@ -1,0 +1,4 @@
+import Waypoint from './waypoint';
+export default Waypoint;
+
+export { waypoint, POSITIONS, getWindow } from './waypoint-hoc';
